The second homework done by Taras Panasiuk.
